updated calendarApp
Changed the way quotes are being displayed. Created a custom dialog and setting modal to false to not block timer events if user isnt around to close out of the dialog box.
Cleanned up timerEvent method by creating two private methods to handle dialog box and isItQuoteTime.
Added a way to prevent user from running application if the EventLinkedList is pointing to an incorrect file.

FIXED BUG that would happen if calendar app was being ran for several days that the application was unable to find the correct current day. Added tempCalendar and manually increment days in timerEvent.